modularity c++
